+++ pssapi/src/main/java/com/example/pssapi/model/entity/Servico.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -15,7 +17,7 @@ public class Servico {
     private Long id;
 
     private String Nome;
-    private float preco;
+    private BigDecimal preco;
     private int duracaoMinutos;
     @ManyToOne
     private Cargo cargo;
